wifi_tg_wp_sms/requirements.txt
flask==2.0.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.0
flask-login==0.5.0
python-dotenv==0.19.2
requests==2.26.0
gunicorn==20.1.0
twilio==7.7.0
python-telegram-bot==13.7
mysqlclient==2.1.0
grafana-api==1.0.7
flask-wtf==1.0.0
